Replace the notion of "user data" to "extra data".
What this change does is makes it so extra data can be allocated at the end of
element's normal malloc(), the size of which is specified when the element is
created. A pointer to the extra data can be retrieved with get_extra_data(),
and the application is free to fill that data with whatever it would like.