dr_flac: optimization to Rice decoding.
This optimization uses a lookup table for checking checking whether or not one of the first 4 bits are set. According to my tests, the first 4 bits are by far the most common positions for the set bit.

At this point dr_flac is now at about 1.25x-1.5x slower than the reference implementation.