dr_fs: Add DR_FS_WIN32_USE_EVENT_MUTEX option which forces the use of the Events API for doing mutual exclusion.
The rationale for this is that the CRITICAL_SECTION API does not deadlock when trying to lock a mutex on the same thread. We want to detect and debug deadlocking errors like this during development so we just use the Events API instead. I'm keeping the CRITICAL_SECTION implementation for the moment because, according to the documentation, it's faster and more efficient.