dr_flac: API CHANGE. Simplify seeking so that negative seeking is no longer used.
The rationale for this change is that it greatly simplifies seeking logic for multiplexed Ogg streams, and probably other containers.

The only public API that has changed is the onSeek callback which now takes a third argument which specifies the origin of the seek: from the start of the stream or relative to it's current position. In addition, the seek offset will never be negative.