dr_audio: Begin work on high level helper APIs.
These APIs supports WAV, FLAC and Vorbis decoding using dr_wav, dr_flac and stb_vorbis respectively. The way it works is that they will be enabled if the relevant libraries are #included before dr_audio.

This commit also adds support for "sounds" which are managed by a "sound world". A Sound object can stream it's audio data using callbacks, or it can store it in the associated voice directly.