dr_flac: Implement a new CRC-16 accumulation method.
This new CRC-16 accumulation technique should hopefully improve
efficiency as a result of keeping it tightly coupled with the
bit streamer. It is more efficient because it performs less data
movements, is always aligned to a byte boundary and is done at the
same time as when the bit streamer's cache is reloaded.

The next commit will be removing the old CRC-16 computations which
should hopefully improve performance.