dr_flac: Experimental work on an optimized f32() read API.
This API is called drflac_read_frames_f32() and reads in blocks of PCM
frames instead of individual samples. This is how all reading in
dr_flac will be in the future.

The implementation of drflac_read_frames_f32() as of this commit is an
experiment in how I might do an SSE optimized implementation later.