dr_wav: Add limits to sample rate, channels and bits per sample.
The idea behind this change is to tighten up validation and better
handle potentially corrupt files. Specifically, I've had a report where
a static analyser doesn't like how the "channels" property is tainted.
I think this is due to it being read from a file, and then later used
in a memory allocation without ever being sanitized with a upper-bound
check. This commit is a potential fix for that issue.

Possible fixes the following GitHub issues:
  * Issue #75
  * Issue #76
  * Issue #79
  * Issue #80
  * Issue #83
  * Issue #84