dr_wav: Add support for inst, cue and acid chunks.
This patch adds support for some extra wav metadata. These chunks are often useful for audio-editing software, software synthesizers and samplers. They are read and stored in the same fashion as the smpl was.

In addition to this, each of these new chunks (and the smpl chunk) now have a bool to say if the chunk was found and correctly read. Previously there was no way to tell if the chunk was valid or not.