dr_wav: Implement a new system for parsing chunks as metadata
This system requires iterating through the chunks twice. The first time to work out the maximum memory needed to store the metadata, and the second time to read the data into a nice format. It does however mean that only one allocation is done, and the data is store contiguously in memory.

The type of metadata is chosen through a bitset. Only the types of metadata that you require are read and allocated.