In flac seeking tests, do not assume PCM frame count fits int
When seeking to random frames, remove an implicit assumption that the
number of PCM frames in the FLAC file under test fits in an int.
(Seeking to frame numbers above INT_MAX is still not tested.)

A side effect of this change is that it fixes a compiler warning about
signed/unsigned comparison (-Wsign-compare) when the tests are compiled
with relatively strict compiler warnings enabled.