dr_flac: Move a stack allocation to the heap for Ogg encapsulation.
This only affects Ogg encapsulated files. Prior to this commit, a ~64K
buffer is allocated on the stack which can result in a stack overflow
on old, low-end or embedded hardware. This has been moved to the heap
to improve compatibility.

Public issue https://github.com/mackron/dr_libs/issues/239