dr_wav: Fix an error when loading AIFF files.
The issue here is that loading would terminate upon finding the SSND
chunk (the equivalent to RIFF's "data" chunk) which would result in
loading failing if the COMM chunk (the equivalent to RIFF's "fmt" chunk)
comes after. In addition, there was a subtle error with the handling of
the "offset" property of the SSND chunk.

Public issue https://github.com/mackron/dr_libs/issues/265