dr_wav: Add a validation check for the data chunk size.
It's possible that the size reported by the data chunk exceeds the size
of the file itself which results in the calculation of the total PCM
frame count being totally incorrect. This commit clamps the data chunk
size to the end of the file to ensure there is no overflow.

Public issue https://github.com/mackron/dr_libs/issues/275