dr_flac: Improvements to drflac_open_and_read_pcm_frames_*() and family.
Prior to this commit, the code would indiscriminately allocate a single
block of memory from `totalPCMFrameCount`. This can result in tiny
malignant FLAC files allocating huge blocks of memory.

I've decided to tighten up this code by dynamically expanding the buffer
as the file is decoded. This code path already exists for cases when
`totalPCMFrameCount` is unknown (set to 0), which means this is also a
code simplification.