dr_wav: Minor rearrangement of some bound checking code.
There are two tables, `coeff1Table` and `coeff2Table`, which are both
indexed by the predictor. In practice these two tables are the same size
so technically these are redundant checks, but when reviewing this code
I myself had to spend more time that I would have liked wrapping my head
around what was going on, so I think the explicitness here will help
with clearing this up.

In addition, I've moved some additional bound checking to closer to the
relevant code that looks up into these tables to hopefully aid in
readability and to make it easier to see that the bound check is
actually being performed.